Store CEnumerator values as BigInteger instead of long
CEnumerator values can be 64bit unsigned integers that do not fit into a Java long.
ISO 9899, Section 6.7.2.2 Enumeration specifiers:

```quote
Each enumerated type shall be compatible with char, a signed integer type, or an
unsigned integer type. The choice of type is implementation-defined, [...]
```

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/make-cenumerator-use-biginteger@47374 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cfa/ast/c/CEnumerator.java
@@ -12,6 +12,7 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 
 import java.io.Serial;
+import java.math.BigInteger;
 import java.util.Objects;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.sosy_lab.cpachecker.cfa.ast.AbstractSimpleDeclaration;
@@ -23,7 +24,7 @@ public final class CEnumerator extends AbstractSimpleDeclaration implements CSim
 
   @Serial private static final long serialVersionUID = -2526725372840523651L;
 
-  private final long value;
+  private final BigInteger value;
   private @Nullable CEnumType enumType;
   private final String qualifiedName;
 
@@ -31,7 +32,7 @@ public final class CEnumerator extends AbstractSimpleDeclaration implements CSim
       final FileLocation pFileLocation,
       final String pName,
       final String pQualifiedName,
-      final long pValue) {
+      final BigInteger pValue) {
     super(pFileLocation, pName);
 
     checkNotNull(pName);
@@ -82,7 +83,7 @@ public final class CEnumerator extends AbstractSimpleDeclaration implements CSim
     return enumType.getCompatibleType();
   }
 
-  public long getValue() {
+  public BigInteger getValue() {
     return value;
   }
